Moved Symbol class to the extended pkg.
